pyproject.toml
[project]
name = "comunity"
version = "0.1.0"
description = "myFirstProject"
authors = [
    {name = "유선혁",email = "tjsgur2399@icloud.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
